# api/openapi.yaml
openapi: 3.0.3
info:
  title: Support Roster API
  description: API for managing support team roster and shift scheduling
  version: 1.0.0
  contact:
    name: Support Team

servers:
  - url: http://localhost:8080/api
    description: Development server

paths:
  /teams:
    get:
      tags:
        - Teams
      summary: Get all teams
      description: Retrieve all support teams with their display order and color configuration
      operationId: getAllTeams
      responses:
        '200':
          description: Successfully retrieved teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'

  /shifts:
    get:
      tags:
        - Shifts
      summary: Get shifts by date
      description: Retrieve all shifts for a specific date, optionally filtered by team
      operationId: getShiftsByDate
      parameters:
        - name: date
          in: query
          required: true
          description: "The date to retrieve shifts for (format: yyyy-MM-dd)"
          schema:
            type: string
            format: date
            example: "2024-01-15"
        - name: teamId
          in: query
          required: false
          description: Filter shifts by team ID
          schema:
            type: string
        - name: timezone
          in: query
          required: false
          description: Timezone for shift time conversion
          schema:
            type: string
            default: "UTC"
            example: "Asia/Shanghai"
      responses:
        '200':
          description: Successfully retrieved shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'

  /shifts/{id}:
    get:
      tags:
        - Shifts
      summary: Get shift by ID
      description: Retrieve detailed information about a specific shift
      operationId: getShiftById
      parameters:
        - name: id
          in: path
          required: true
          description: Shift ID
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved shift
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftDetail'
        '404':
          description: Shift not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /staff/{id}:
    get:
      tags:
        - Staff
      summary: Get staff by ID
      description: Retrieve detailed information about a staff member
      operationId: getStaffById
      parameters:
        - name: id
          in: path
          required: true
          description: Staff ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved staff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '404':
          description: Staff not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /staff:
    get:
      tags:
        - Staff
      summary: Get all staff
      description: Retrieve all staff members
      operationId: getAllStaff
      responses:
        '200':
          description: Successfully retrieved staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Staff'

  /role-groups:
    get:
      tags:
        - RoleGroups
      summary: Get all role groups
      description: Retrieve all role groups (teams) from the roster
      operationId: getAllRoleGroups
      responses:
        '200':
          description: Successfully retrieved role groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleGroup'

  /shift-codes:
    get:
      tags:
        - ShiftCodes
      summary: Get all shift codes
      description: Retrieve all shift code definitions with colors
      operationId: getAllShiftCodes
      responses:
        '200':
          description: Successfully retrieved shift codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftCode'

components:
  schemas:
    Team:
      type: object
      description: Support team information
      properties:
        id:
          type: string
          description: Unique team identifier
          example: "incident-manager"
        name:
          type: string
          description: Display name of the team
          example: "Incident Manager"
        color:
          type: string
          description: Color theme for the team (blue, green, orange, purple, red)
          enum: [blue, green, orange, purple, red]
          example: "orange"
        order:
          type: integer
          description: Display order
          example: 0
      required:
        - id
        - name
        - color
        - order

    Shift:
      type: object
      description: Shift information for timeline display
      properties:
        id:
          type: string
          description: Unique shift identifier
          example: "s1"
        teamId:
          type: string
          description: Team ID this shift belongs to
          example: "incident-manager"
        staffId:
          type: integer
          format: int64
          description: Staff member ID
          example: 123
        userName:
          type: string
          description: Staff member name
          example: "Alex Chen"
        userAvatar:
          type: string
          description: URL to staff member avatar
          example: "https://example.com/avatar.jpg"
        code:
          type: string
          description: Shift code (A, B, C, D, DS, NS, OC, BH, HoL)
          example: "OC"
        meaning:
          type: string
          description: Shift meaning/description
          example: "Full Day Oncall Support"
        start:
          type: string
          format: date-time
          description: Shift start time in ISO format
          example: "2024-01-15T00:00:00Z"
        end:
          type: string
          format: date-time
          description: Shift end time in ISO format
          example: "2024-01-15T23:59:00Z"
        timezone:
          type: string
          description: Timezone of the shift
          example: "HKT"
        isPrimary:
          type: boolean
          description: Whether this is a primary on-call shift
          example: true
        showOnRoster:
          type: boolean
          description: Whether to show on roster page
          example: true
        remark:
          type: string
          description: Additional remarks
          example: "Primary Incident Manager on the duty week"
        contact:
          $ref: '#/components/schemas/Contact'
        backup:
          $ref: '#/components/schemas/Backup'
      required:
        - id
        - teamId
        - staffId
        - userName
        - code
        - start
        - end
        - timezone

    ShiftDetail:
      allOf:
        - $ref: '#/components/schemas/Shift'
        - type: object
          properties:
            color:
              type: string
              description: Hex color code for the shift
              example: "#FF6347"

    Contact:
      type: object
      description: Contact information for staff member
      properties:
        slack:
          type: string
          description: Slack handle
          example: "@achen"
        email:
          type: string
          description: Email address
          example: "achen@company.com"
        phone:
          type: string
          description: Phone number
          example: "+1-555-0101"
      required:
        - slack
        - email
        - phone

    Backup:
      type: object
      description: Backup staff information
      properties:
        name:
          type: string
          description: Backup staff name
          example: "Alex Chen"
        contact:
          type: string
          description: Backup staff contact (Slack handle)
          example: "@achen"

    Staff:
      type: object
      description: Staff member information
      properties:
        id:
          type: integer
          format: int64
          description: Unique staff identifier
          example: 123
        name:
          type: string
          description: Full name
          example: "Alex Chen"
        avatar:
          type: string
          description: Avatar URL
          example: "https://example.com/avatar.jpg"
        email:
          type: string
          description: Email address
          example: "achen@company.com"
        phone:
          type: string
          description: Phone number
          example: "+1-555-0101"
        slack:
          type: string
          description: Slack handle
          example: "@achen"
        roleGroups:
          type: array
          description: Role groups the staff belongs to
          items:
            type: string
          example: ["L1_China", "Incident_Manager_China"]
      required:
        - id
        - name

    RoleGroup:
      type: object
      description: Role group (team) information
      properties:
        id:
          type: string
          description: Role group identifier
          example: "L1_China"
        name:
          type: string
          description: Display name
          example: "L1 China"
        category:
          type: string
          description: Category (L1, L2, L2+, L3, Incident_Manager, DevOps)
          example: "L1"
        region:
          type: string
          description: Region (China, India, AP, EMEA, MDP)
          example: "China"
      required:
        - id
        - name

    ShiftCode:
      type: object
      description: Shift code definition with color
      properties:
        code:
          type: string
          description: Shift code
          example: "A"
        meaning:
          type: string
          description: Shift meaning
          example: "00:00-07:00"
        colorName:
          type: string
          description: Color name
          example: "Orange"
        colorHex:
          type: string
          description: Hex color code
          example: "#FFA500"
      required:
        - code
        - meaning

    ErrorResponse:
      type: object
      description: Error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        path:
          type: string
          description: Request path
      required:
        - timestamp
        - status
        - error
        - message