# src/main/java/com/support/server/supportrosterserver/dto/ShiftCodeDto.java
package com.support.server.supportrosterserver.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class ShiftCodeDto {
    private String code;
    private String meaning;
    private String colorName;
    private String colorHex;
}