# src/main/java/com/support/server/supportrosterserver/dto/TeamDto.java
package com.support.server.supportrosterserver.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TeamDto {
    private String id;
    private String name;
    private String color;
    private Integer order;
}